Create README.md
This application represents a simple student project. The task of this application is to enable easier recording of art works in art galleries. This application is primarily intended for art gallery employees. After running the application, a login window opens where the employee enters the necessary login data (note: the employee must be in the art gallery employee database). The employee can view the information about the art works in the gallery. Also, the employee can edit the information about the existing works, and add art works in the art gallery or delete art works from the gallery. Also, after running the application, the login window offers the option ‘log in as a guest’. This option allows art gallery visitors to view which artworks are in the gallery and to find out some information about them. Art gallery visitors do not have the possibility to delete or add an art work. Also, art gallery visitors do not have the possibility to edit information about the existing art works . This application can easily be improved by adding additional features.